+++ packages/core/src/components/templates/FieldHelpTemplate.tsx
@@ -1,4 +1,5 @@
-import { helpId, FieldHelpProps, FormContextType, RJSFSchema, StrictRJSFSchema } from '@rjsf/utils';
+import { helpId, FieldHelpProps, FormContextType, RJSFSchema, StrictRJSFSchema, getUiOptions } from '@rjsf/utils';
+import Markdown from 'markdown-to-jsx';
 
 /** The `FieldHelpTemplate` component renders any help desired for a field
  *
@@ -9,12 +10,20 @@ export default function FieldHelpTemplate<
   S extends StrictRJSFSchema = RJSFSchema,
   F extends FormContextType = any,
 >(props: FieldHelpProps<T, S, F>) {
-  const { fieldPathId, help } = props;
+  const { fieldPathId, help, uiSchema = {}, registry } = props;
   if (!help) {
     return null;
   }
   const id = helpId(fieldPathId);
+  const uiOptions = getUiOptions<T, S, F>(uiSchema, registry?.globalUiOptions);
   if (typeof help === 'string') {
+    if (uiOptions.enableMarkdownInHelp) {
+      return (
+        <div id={id} className='help-block'>
+          <Markdown options={{ disableParsingRawHTML: true }}>{help}</Markdown>
+        </div>
+      );
+    }
     return (
       <p id={id} className='help-block'>
         {help}