+++ packages/primereact/src/FieldHelpTemplate/FieldHelpTemplate.tsx
@@ -1,4 +1,5 @@
-import { FieldHelpProps, FormContextType, RJSFSchema, StrictRJSFSchema, helpId } from '@rjsf/utils';
+import { FieldHelpProps, FormContextType, RJSFSchema, StrictRJSFSchema, helpId, getUiOptions } from '@rjsf/utils';
+import Markdown from 'markdown-to-jsx';
 
 /** The `FieldHelpTemplate` component renders any help desired for a field
  *
@@ -9,9 +10,17 @@ export default function FieldHelpTemplate<
   S extends StrictRJSFSchema = RJSFSchema,
   F extends FormContextType = any,
 >(props: FieldHelpProps<T, S, F>) {
-  const { fieldPathId, help } = props;
+  const { fieldPathId, help, uiSchema = {}, registry } = props;
   if (help) {
     const id = helpId(fieldPathId);
+    const uiOptions = getUiOptions<T, S, F>(uiSchema, registry?.globalUiOptions);
+    if (typeof help === 'string' && uiOptions.enableMarkdownInHelp) {
+      return (
+        <small id={id}>
+          <Markdown options={{ disableParsingRawHTML: true }}>{help}</Markdown>
+        </small>
+      );
+    }
     return <small id={id}>{help}</small>;
   }
   return null;