+++ CHANGELOG.md
@@ -26,6 +26,7 @@ should change the heading of the (upcoming) version to include a major version b
 
 - Modified `CheckboxesWidget` to render the Title, fixing ([#4840](https://github.com/rjsf-team/react-jsonschema-form/issues/4840))
 - Updated `CheckboxWidget` to handle label and description rendering consistently, fixing ([#4742](https://github.com/rjsf-team/react-jsonschema-form/issues/4742))
+- Updated `FieldHelpTemplate` to render markdown when `ui:enableMarkdownInHelp` is true
 
 ## @rjsf/core
 
@@ -38,10 +39,16 @@ should change the heading of the (upcoming) version to include a major version b
 ## @rjsf/fluentui-rc
 
 - Updated `FieldTemplate` to skip label and description rendering for checkbox widgets, fixing ([#4742](https://github.com/rjsf-team/react-jsonschema-form/issues/4742))
+- Updated `FieldHelpTemplate` to render markdown when `ui:enableMarkdownInHelp` is true
+
+## @rjsf/daisyui
+
+- Updated `FieldHelpTemplate` to render markdown when `ui:enableMarkdownInHelp` is true
 
 ## @rjsf/mantine
 
 - Updated `CheckboxWidget` to handle label and description rendering consistently, fixing ([#4742](https://github.com/rjsf-team/react-jsonschema-form/issues/4742))
+- Updated `FieldHelpTemplate` to render markdown when `ui:enableMarkdownInHelp` is true
 
 ## @rjsf/mui
 
@@ -51,19 +58,23 @@ should change the heading of the (upcoming) version to include a major version b
 ## @rjsf/primereact
 
 - Updated `FieldTemplate` to skip label and description rendering for checkbox widgets, fixing ([#4742](https://github.com/rjsf-team/react-jsonschema-form/issues/4742))
+- Updated `FieldHelpTemplate` to render markdown when `ui:enableMarkdownInHelp` is true
 
 ## @rjsf/react-bootstrap
 
 - Updated `FieldTemplate` to skip label and description rendering for checkbox widgets, fixing ([#4742](https://github.com/rjsf-team/react-jsonschema-form/issues/4742))
+- Updated `FieldHelpTemplate` to render markdown when `ui:enableMarkdownInHelp` is true
 
 ## @rjsf/semantic-ui
 
 - Updated `FieldTemplate` to skip label and description rendering for checkbox widgets, fixing ([#4742](https://github.com/rjsf-team/react-jsonschema-form/issues/4742))
+- Updated `FieldHelpTemplate` to render markdown when `ui:enableMarkdownInHelp` is true
 
 ## @rjsf/shadcn
 
 - Updated `FieldTemplate` to skip label and description rendering for checkbox widgets, fixing ([#4742](https://github.com/rjsf-team/react-jsonschema-form/issues/4742))
 - Updated the `Command` component to properly handle `forwardRef`
+- Updated `FieldHelpTemplate` to render markdown when `ui:enableMarkdownInHelp` is true
 
 ## @rjsf/utils